Prepare matrices with default patterns and unmasked data A list with a matrix containing the default patterns (finder pattern, timing pattern, separators, alignment pattern and dark module), the unmask data pattern and the version.

Prepare matrices with default patterns and unmasked data A list with a matrix containing the default patterns (finder pattern, timing pattern, separators, alignment pattern and dark module), the unmask data pattern and the version.

qr_matrix(x, ecl = c("L", "M", "Q", "H"))

Arguments

x

the input string

ecl

the required error correction level. Available options are "L" (7%), "M" (15%), "Q" (25%) and "H" (30%). Defaults to "L".

See also

Author

Thierry Onkelinx

Examples

qr_matrix("HELLO WORLD")
#> $patterns
#>        [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8] [,9] [,10] [,11] [,12]
#>  [1,]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE   NA    NA    NA    NA
#>  [2,]  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE   NA    NA    NA    NA
#>  [3,]  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE   NA    NA    NA    NA
#>  [4,]  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE   NA    NA    NA    NA
#>  [5,]  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE   NA    NA    NA    NA
#>  [6,]  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE   NA    NA    NA    NA
#>  [7,]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE TRUE FALSE  TRUE FALSE
#>  [8,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE   NA    NA    NA    NA
#>  [9,]    NA    NA    NA    NA    NA    NA  TRUE    NA   NA    NA    NA    NA
#> [10,]    NA    NA    NA    NA    NA    NA FALSE    NA   NA    NA    NA    NA
#> [11,]    NA    NA    NA    NA    NA    NA  TRUE    NA   NA    NA    NA    NA
#> [12,]    NA    NA    NA    NA    NA    NA FALSE    NA   NA    NA    NA    NA
#> [13,]    NA    NA    NA    NA    NA    NA  TRUE    NA   NA    NA    NA    NA
#> [14,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE    NA    NA    NA
#> [15,]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE   NA    NA    NA    NA
#> [16,]  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE   NA    NA    NA    NA
#> [17,]  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE   NA    NA    NA    NA
#> [18,]  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE   NA    NA    NA    NA
#> [19,]  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE FALSE   NA    NA    NA    NA
#> [20,]  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE   NA    NA    NA    NA
#> [21,]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE   NA    NA    NA    NA
#>       [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,]    NA FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [2,]    NA FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [3,]    NA FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE
#>  [4,]    NA FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE
#>  [5,]    NA FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE
#>  [6,]    NA FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [7,]  TRUE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [8,]    NA FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [9,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [10,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [11,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [12,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [13,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [14,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [15,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [16,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [17,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [18,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [19,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [20,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> [21,]    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> 
#> $payload
#>        [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10] [,11] [,12]
#>  [1,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [2,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE
#>  [3,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [4,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE
#>  [5,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE
#>  [6,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#>  [7,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [8,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE
#>  [9,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [10,]  TRUE FALSE  TRUE  TRUE  TRUE FALSE FALSE  TRUE  TRUE  TRUE  TRUE FALSE
#> [11,] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE  TRUE
#> [12,]  TRUE  TRUE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [13,]  TRUE FALSE  TRUE  TRUE FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE
#> [14,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE
#> [15,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [16,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE
#> [17,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE
#> [18,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE
#> [19,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#> [20,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [21,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>       [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#>  [1,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [2,]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [3,]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [4,]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [5,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [6,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [7,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [8,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [9,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [10,] FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE
#> [11,]  TRUE FALSE FALSE FALSE FALSE  TRUE  TRUE FALSE  TRUE
#> [12,]  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [13,]  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [14,] FALSE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE
#> [15,] FALSE FALSE  TRUE  TRUE  TRUE  TRUE FALSE FALSE  TRUE
#> [16,] FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [17,] FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE
#> [18,] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#> [19,]  TRUE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [20,]  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE
#> [21,]  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE FALSE FALSE
#> 
#> $version
#> [1] 1
#>