Determine the required mode

The current implementation handles three modes: numeric, alphanumeric and byte. Kanji is currently not supported. Please contact the maintainer if you need it. Numeric: only digits from 0 to 9 Alphanumeric: all numeric characters, upper case LETTERS, and the characters " " (space), "$", "%", "*", "+", "-", ".", "/" and ":" Byte: All characters from the Latin 1 (ISO 8859-1) character set. Input strings with an other encoding are converted into Latin 1. The function return an error if such conversion fails.

Usage

qr_mode(x)

Arguments

x

the input string

Value

a character indicating the mode

See also

Author

Thierry Onkelinx

Examples

qr_mode("0123")
#> [1] "Numeric"
qr_mode("A")
#> [1] "Alphanumeric"
qr_mode("a")
#> [1] "Byte"