Create a QR code for a vCard

Create a QR code for a vCard

Usage

qr_vcard(
  given,
  family,
  address,
  email,
  telephone,
  organisation,
  job_title,
  url,
  gender,
  logo,
  photo,
  middle = character(0),
  prefix = character(0),
  suffix = character(0),
  ecl = c("L", "M", "Q", "H"),
  ...
)

Arguments

given

The given name.

family

The family name.

address

In case of a single address, a named character vector with the following elements: street_nr, city, region, postal_code and country. In case of multiple addresses, a named list of named character vectors. The names of the list are used as the type of the address.

email

Optionally one or more email addresses. The names of the vector are used as the type of the email address.

telephone

Optionally one of more telephone numbers. The names of the vector are used as the type of the telephone number.

organisation

Optionally the name of your organisation and team within the organisation.

job_title

Optionally the job title of the person.

url

Optionally one or more URLs. The names of the vector are used as the type of the URL.

gender

Optionally a string describing the gender of the person.

logo

Optionally a URL to a logo.

photo

Optionally a URL to a photo.

middle

Optionally one or more middle names.

prefix

Optionally one or more prefixes.

suffix

Optionally one or more suffixes.

ecl

the required error correction level. Available options are "L" (7%), "M" (15%), "Q" (25%) and "H" (30%). Defaults to "L".

...

Additional arguments are silently ignored.